fix(conversation):COCO-4213 mark as read from message navigation
